Move: google methods to `lib` folder
